requirements.txt
gunicorn
django>=4.2,<5.3
djangorestframework>=3.14,<3.16
psycopg2
drf-spectacular>=0.27,<0.29